Преобразование в дробные числа,дробный калькулятор

+++ py/hello.py
@@ -1,3 +1,3 @@
-a = int(input())
-b = int(input())
+a = float(input())
+b = float(input())
 print("Сумма=" + str(a+b))
